[Rename] collectedActions.ts -> collectActions.ts
