--- types/zustand.d.ts
@@ -1,2 +1,0 @@
-// declare module 'zustand';
-// declare module 'zustand/middleware';
